+++ Database/ClassifiedsDatabase/Scripts-3.31/001-EventBooking.HowYouHeardAboutEvent.sql
@@ -1,0 +1,2 @@
+﻿ALTER TABLE EventBooking
+ADD HowYouHeardAboutEvent VARCHAR(100) NULL